+++ src/codeactions/alCodeActionsProvider.ts
@@ -45,17 +45,25 @@ export class ALCodeActionsProvider implements vscode.CodeActionProvider {
             new ALSortPermissionSetListCommand(this._toolsExtensionContext),
             
             new ALCreateInterfaceCodeCommand(this._toolsExtensionContext),
+            
             //diagnostics fixes
-            //AA0005 fix disabled, needs some fixes before going live
-            //new ALCodeCopFixAA0005(this._toolsExtensionContext),
             new ALCodeCopFixAA0008(this._toolsExtensionContext),
             new ALCodeCopFixAA0137(this._toolsExtensionContext),
             new ALCodeCopFixAA0139(this._toolsExtensionContext)];
     }
 
     provideCodeActions(document: vscode.TextDocument, range: vscode.Range | vscode.Selection, context: vscode.CodeActionContext, token: vscode.CancellationToken): vscode.ProviderResult<(vscode.Command | vscode.CodeAction)[]> {
-        let diag = vscode.languages.getDiagnostics(document.uri);        
-        return this.collectCodeActions(document,range, diag);
+        //load diagnostics only if CodeCop fixes are enabled
+        let settings =  vscode.workspace.getConfiguration('alOutline', document.uri);
+        let enableCodeCopFixes = !!settings.get<boolean>('enableCodeCopFixes');
+        let diagnostics: vscode.Diagnostic[];
+        if (enableCodeCopFixes)
+            diagnostics = vscode.languages.getDiagnostics(document.uri);
+        else
+            diagnostics = [];
+        
+        //collect code actions
+        return this.collectCodeActions(document,range, diagnostics);
     }
 
     protected async collectCodeActions(document: vscode.TextDocument, range: vscode.Range | vscode.Selection, diagnostic: vscode.Diagnostic[]): Promise<vscode.CodeAction[]> {